cat_dog_app/requirements.txt
tensorflow>=2.12.0
opencv-python>=4.5.5.64
Pillow>=9.0.0
numpy>=1.21.0
